# Conflicts:
#	public/css/app.css
#	public/css/app.css.map
#	public/css/freyja.css
#	public/css/freyja.css.map
#	public/css/front.css
#	public/css/front.css.map
#	public/css/map-v3.css
#	public/css/map-v3.css.map
#	public/js/app.js
#	public/js/app.js.map
#	public/js/attributes.js
#	public/js/attributes.js.map
#	public/js/forms/calendar.js
#	public/js/forms/calendar.js.map
#	public/js/location/map.js
#	public/js/location/map.js.map
#	public/mix-manifest.json

+++ app/Http/Controllers/Maps/MapMarkerController.php
@@ -118,6 +118,9 @@ class MapMarkerController extends Controller
     public function edit(Map $map, MapMarker $mapMarker)
     {
         $this->authorize('update', $map);
+        if ($mapMarker->map_id !== $map->id) {
+            abort(503);
+        }
 
         $ajax = request()->ajax();
         $model = $mapMarker;
@@ -140,6 +143,9 @@ class MapMarkerController extends Controller
     public function update(StoreMapMarker $request, Map $map, MapMarker $mapMarker)
     {
         $this->authorize('update', $map);
+        if ($mapMarker->map_id !== $map->id) {
+            abort(503);
+        }
 
         // For ajax requests, send back that the validation succeeded, so we can really send the form to be saved.
         if (request()->ajax()) {
@@ -176,6 +182,9 @@ class MapMarkerController extends Controller
     public function destroy(Map $map, MapMarker $mapMarker)
     {
         $this->authorize('update', $map);
+        if ($mapMarker->map_id !== $map->id) {
+            abort(503);
+        }
 
         $mapMarker->delete();
